Add args to backup script to control delete and time

+++ python/backup.py
@@ -1,8 +1,8 @@
-"""Simple, same-site backup.
+"""Simple, same-site backup via Zip.
 
 The MIT License (MIT)
 
-Copyright (c) 2014 Joshua S. Ziegler 
+Copyright (c) 2014 Joshua S. Ziegler
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -28,27 +28,33 @@ _log = logging.getLogger(__name__) # get module-level logger
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument('--src_path',    type=str, default="C:/home/")
-    parser.add_argument('--dst_path',    type=str, default="D:/") 
-    parser.add_argument('--name_prefix', type=str, default="josh_pc_") 
+    parser.add_argument('--src_path',    type=str, default="D:/home/")
+    parser.add_argument('--dst_path',    type=str, default="D:/")
+    parser.add_argument('--name_prefix', type=str, default="josh_pc_")
+    parser.add_argument('-d', '--delete_old',  action='store_true')
+    parser.add_argument('-t', '--append_time', action='store_true')
     args = parser.parse_args()
 
     # Don't edit below here
     now = datetime.datetime.now()
-    archive_name = args.name_prefix + now.isoformat().replace(":", '_').replace(".","_").replace("-","_")
+    date_format = "%Y_%m_%d"
+    if args.append_time:
+        date_format += "-%H_%M_%S"
+
+    archive_name = args.name_prefix + now.strftime(date_format)
     dst_new = os.path.join(args.dst_path,  archive_name)
     format = '%(levelname)s:%(filename)s:%(lineno)s:%(funcName)s:  %(message)s'
     logging.basicConfig(format=format, level=logging.INFO)
 
-    # find old backups first so we don't delete our new one
-    old_backups = [ f for f in os.listdir(args.dst_path) if os.path.isfile(os.path.join(args.dst_path,f)) and args.name_prefix in f ]
+    if args.delete_old:
+        # find old backups with the same name_prefix
+        old_backups = [ f for f in os.listdir(args.dst_path) if os.path.isfile(os.path.join(args.dst_path,f)) and args.name_prefix in f ]
+        # delete the old backups
+        for path in old_backups:
+            tmp = os.path.join(args.dst_path, path)
+            _log.info("Deleting old backups: {}".format(tmp))
+            os.remove(tmp)
 
     # create the new backup
     _log.info("Creating new backup: {}".format(dst_new))
     shutil.make_archive(dst_new, "zip", args.src_path, args.src_path)
-
-    # delete the old backups
-    for path in old_backups:
-        tmp = os.path.join(args.dst_path, path)
-        _log.info("Deleting old backups: {}".format(tmp))
-        os.remove(tmp)
